requirements.txt
annotated-types==0.6.0
anthropic==0.25.7
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.14.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
openai==1.26.0
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1

mailbox
requests

nltk
spaCy
pandas
pytest
# chromadb_client
# langchain
# chromadb

# data extraction
google-api-python-client
pypdf2
openpyxl
python-docx

# vector database
pinecone
tiktoken

# streamlit
streamlit
streamlit-extras
streamlit-option-menu
weasyprint

# Google API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

docx2txt
xlrd
bs4

# dedupe_sqlite_python.py
numpy
# picklestreamlit-extras
